requirements.txt
python-dotenv
crewai
crewai_tools
rich
agentops
openai